Guard against domain errors in OD dithering
When OD dithering is turned on, a very slow-growing culture can have zero (or negative) OD when the OD control function "Zigzag" is called. This triggers a domain error when calculating a fold-change since the last OD measurement, which crashes the control function. 

This fix checks whether either the current or last OD measurements were nonpositive. If either one is, it does not attempt to calculate a growth rate.