Update app.py
This is a bit of an obscure fix. 
After a lot of testing, it seems that on the V1.2 reactors there is some possibility of pump communication errors occuring and leading to freezing of  the I2C bus for ~1-2 seconds, which can lead to much more liquid pumped than intended! 
It seems this communication error almost exclusively happens when READING from the I2C bus - which is only done in the setPWM function when we are doing our check to make sure the values written to the device were received as expected. Ironically, it therefore seems this safety check is causing our errors! Further, it seems these errors are exacerbated by having the stirring enabled at the same time as we are reading from the pump board.

Therefore, I have changed setPWM so that it only reads back the PWM for other variables (e.g. the LEDs/heater/stirrer), and for the pumps it now sends commands twice so (hopefully) in the rare occasion that there is any issue, the Pump is hit again with the correct command set.